Ensure translations are html escaped
Reviewing translated content may or may not always be feasible. With that in
mind, we may apply a protective coat of paranoia by always html-escaping
translated strings when they are passed to html-parsing functions like html()
and append(). This ensures that if any malicious html makes its way into
translation files, they are rendered safely in text nodes.